chore: clippy auto fix

+++ async-std/src/server/unix.rs
@@ -86,13 +86,12 @@ impl Server for AsyncStdServer {
     }
 
     async fn clean_up(self) {
-        if let Unix(u) = &self.0 {
-            if let Ok(local) = u.local_addr() {
-                if let Some(path) = local.as_pathname() {
-                    log::info!("deleting {:?}", &path);
-                    log_error!(async_std::fs::remove_file(path).await);
-                }
-            }
+        if let Unix(u) = &self.0
+            && let Ok(local) = u.local_addr()
+            && let Some(path) = local.as_pathname()
+        {
+            log::info!("deleting {:?}", &path);
+            log_error!(async_std::fs::remove_file(path).await);
         }
     }
 }
