+++ smol/src/server/unix.rs
@@ -66,13 +66,12 @@ impl Server for SmolServer {
     }
 
     async fn clean_up(self) {
-        if let Unix(u) = &self.0 {
-            if let Ok(local) = u.local_addr() {
-                if let Some(path) = local.as_pathname() {
-                    log::info!("deleting {:?}", &path);
-                    log_error!(std::fs::remove_file(path));
-                }
-            }
+        if let Unix(u) = &self.0
+            && let Ok(local) = u.local_addr()
+            && let Some(path) = local.as_pathname()
+        {
+            log::info!("deleting {:?}", &path);
+            log_error!(std::fs::remove_file(path));
         }
     }
 }
